[#33]refactor: 카드 크기 원래대로 수정, 함수파일 이름 변경
