+++ components/commons/card/Card.module.scss
@@ -1,5 +1,5 @@
 .cardContainer {
-  width: 100%;
+  width: 312px;
   height: 349px;
   display: inline-flex;
   padding: 16px;
